GP-950 refactor admin structure
